Begin removing agency lat/lng
