+++ db/helpers_/result_formatting.py
@@ -62,8 +62,8 @@ def agency_to_agency_dict(agency: Agency) -> dict[str, Any]:
         "submitted_name": agency.name,
         "name": agency.name,
         "homepage_url": agency.homepage_url,
-        "lat": agency.lat,
-        "lng": agency.lng,
+        "lat": None,
+        "lng": None,
         "defunct_year": agency.defunct_year,
         "agency_type": agency.agency_type,
         "multi_agency": agency.multi_agency,
