+++ endpoints/instantiations/agencies_/put/schemas/inner.py
@@ -1,4 +1,4 @@
-from middleware.schema_and_dto.dtos.agencies.put import AgencyInfoPutDTO
+from endpoints.instantiations.agencies_.put.dto import AgencyInfoPutDTO
 from middleware.schema_and_dto.dynamic.pydantic_to_marshmallow.core import (
     pydantic_to_marshmallow,
 )
