+++ endpoints/schema_config/instantiations/agencies/by_id/put.py
@@ -1,5 +1,5 @@
 from endpoints.schema_config.helpers import get_put_resource_endpoint_schema_config
-from middleware.schema_and_dto.schemas.agencies.put import AgenciesPutSchema
+from endpoints.instantiations.agencies_.put.schemas.outer import AgenciesPutSchema
 
 AgenciesByIDPutEndpointSchemaConfig = get_put_resource_endpoint_schema_config(
     input_schema=AgenciesPutSchema(),
